@@ -2,7 +2,7 @@ package com.newage.letstalk.activity;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.widget.Toast;
 
 import com.newage.letstalk.SessionManager;
