@@ -3,7 +3,7 @@ package com.newage.letstalk.activity;
 import android.content.Intent;
 import android.graphics.PixelFormat;
 import android.os.Handler;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.Window;
 import android.view.animation.Animation;
