@@ -1,13 +1,11 @@
 package com.newage.letstalk.utils;
 
 import android.app.Activity;
-import android.app.Fragment;
 import android.content.Context;
-import android.content.pm.PackageManager;
 import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
-import android.support.v4.app.ActivityCompat;
-import android.support.v4.content.ContextCompat;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
 import android.view.View;
 import android.view.inputmethod.InputMethodManager;
 
@@ -16,10 +14,8 @@ import com.google.i18n.phonenumbers.Phonenumber;
 
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
-import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Date;
-import java.util.List;
 import java.util.Locale;
 import java.util.Objects;
 
