@@ -1,14 +1,13 @@
 package com.newage.letstalk.dataLayer.local;
 
-import android.arch.persistence.room.Database;
-import android.arch.persistence.room.Room;
-import android.arch.persistence.room.RoomDatabase;
-import android.arch.persistence.room.TypeConverters;
+import androidx.room.Database;
+import androidx.room.Room;
+import androidx.room.RoomDatabase;
+import androidx.room.TypeConverters;
 import android.content.Context;
 
 import com.newage.letstalk.dataLayer.local.dao.FriendDao;
 import com.newage.letstalk.dataLayer.local.tables.Friend;
-
 
 @Database(entities = {Friend.class}, version = 1, exportSchema = false)
 @TypeConverters({Converters.class})
