@@ -1,7 +1,6 @@
 package com.newage.letstalk.dataLayer.local;
 
-
-import android.arch.persistence.room.TypeConverter;
+import androidx.room.TypeConverter;
 
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
@@ -10,7 +9,6 @@ import java.lang.reflect.Type;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
-
 
 public class Converters {
     @TypeConverter
