@@ -1,16 +1,15 @@
 package com.newage.letstalk.dataLayer.local.dao;
 
-import android.arch.lifecycle.LiveData;
-import android.arch.persistence.room.Dao;
-import android.arch.persistence.room.Insert;
-import android.arch.persistence.room.OnConflictStrategy;
-import android.arch.persistence.room.Query;
-import android.arch.persistence.room.Update;
+import androidx.lifecycle.LiveData;
+import androidx.room.Dao;
+import androidx.room.Insert;
+import androidx.room.OnConflictStrategy;
+import androidx.room.Query;
+import androidx.room.Update;
 
 import com.newage.letstalk.dataLayer.local.tables.Friend;
 
 import java.util.List;
-
 
 @Dao
 public interface FriendDao {
@@ -47,4 +46,5 @@ public interface FriendDao {
 
     @Query("DELETE FROM friends")
     void deleteAllFriends();
+
 }
